pools: improve rebalanced base pool tests
