+++ pkg/asset-manager-utils/contracts/test/MockAssetManager.sol
@@ -24,6 +24,7 @@ contract MockAssetManager is IAssetManager {
     event Rebalanced(address assetManager, bytes32 poolId, IERC20 token, bool force);
 
     IERC20 internal _token;
+    bool public shouldRebalanceMocked;
 
     constructor(IERC20 token) {
         _token = token;
@@ -53,11 +54,15 @@ contract MockAssetManager is IAssetManager {
         // solhint-disable-previous-line no-empty-blocks
     }
 
-    function shouldRebalance(uint256, uint256) external pure override returns (bool) {
-        return true;
+    function shouldRebalance(uint256, uint256) external view override returns (bool) {
+        return shouldRebalanceMocked;
     }
 
     function rebalance(bytes32 poolId, bool force) external override {
         emit Rebalanced(address(this), poolId, _token, force);
+    }
+
+    function mockShouldRebalance(bool newShouldRebalance) external {
+        shouldRebalanceMocked = newShouldRebalance;
     }
 }
